중복 폴더 제거 , 파일 위치 수정
