+++ backend/grids/src/main/java/com/grids/domain/item/controller/ItemController.java
@@ -1,22 +1,26 @@
 package com.grids.domain.item.controller;
 
+import com.grids.domain.item.dto.ItemInfoUpdateRequestDto;
+import com.grids.domain.item.dto.ItemInfoUpdateResponseDto;
 import com.grids.domain.item.dto.ItemListDto;
 import com.grids.domain.item.entity.Item;
 import com.grids.domain.item.service.ItemService;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.ResponseEntity;
 import org.springframework.transaction.annotation.Transactional;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
 
+
 @RestController
+@RequestMapping("/items")
 @RequiredArgsConstructor
 public class ItemController {
 
     private final ItemService itemService;
 
-    @GetMapping("/items")
+    @GetMapping
     @Transactional(readOnly = true)
     public List<ItemListDto> list() {
         List<ItemListDto> itemList =  itemService.findAll().stream()
@@ -26,5 +30,10 @@ public class ItemController {
         return itemList;
     }
 
+    @PatchMapping("/{itemId}")
+    public ResponseEntity<ItemInfoUpdateResponseDto> updateItemInfo(@PathVariable("itemId") Long itemId, @RequestBody ItemInfoUpdateRequestDto itemInfoUpdateRequestDto) {
+        ItemInfoUpdateResponseDto responseDto = itemService.updateItem(itemId, itemInfoUpdateRequestDto);
+        return ResponseEntity.ok(responseDto);
+    }
 
 }
